+++ R/separate.R
@@ -94,7 +94,7 @@ separate_.data.frame <- function(data, col, into, sep = "[^[:alnum:]]+",
 
   names(l) <- into
   if (convert) {
-    l[] <- lapply(l, type.convert, as.is = FALSE)
+    l[] <- lapply(l, type.convert, as.is = TRUE)
   }
 
   # Insert into existing data frame