+++ NEWS.md
@@ -38,6 +38,9 @@
   base R regular expressions. It also returns NA instead of throwing
   an error if the regular expression doesn't match (#72).
   
+* `extract()` and `separate()` preserve character vectors when
+  `covert` is TRUE (#99).
+  
 * The internals of `spread()` have been rewritten, and now preserve all 
   attributes of the input `value` column. This means that you can now 
   spread date (#62) and factor (#35) inputs.